Move Draw2d interface back to IDraw2d
