+++ Gems/LyShine/Code/Include/LyShine/Draw2d.h
@@ -8,13 +8,10 @@
 #pragma once
 
 #include <LyShine/IDraw2d.h>
-#include <LyShine/ILyShine.h>
-#include <LyShine/Bus/UiTransformBus.h>
 
 #include <AzFramework/Font/FontInterface.h>
 #include <Atom/Bootstrap/BootstrapNotificationBus.h>
 #include <Atom/RPI.Public/DynamicDraw/DynamicDrawInterface.h>
-#include <Atom/RPI.Reflect/Image/Image.h>
 #include <Atom/RPI.Public/ViewportContext.h>
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -23,68 +20,10 @@
 //! The CDraw2d class implements the IDraw2d interface for drawing 2D images, shapes and text.
 //! Positions and sizes are specified in pixels in the associated 2D viewport.
 class CDraw2d
-    : public IDraw2d // [LYSHINE_ATOM_TODO][GHI #3573] Make Draw2d work better as an API
+    : public IDraw2d
     , public AZ::Render::Bootstrap::NotificationBus::Handler
 {
 public: // types
-
-    struct RenderState
-    {
-        RenderState()
-        {
-            m_blendState.m_enable = true;
-            m_blendState.m_blendSource = AZ::RHI::BlendFactor::AlphaSource;
-            m_blendState.m_blendDest = AZ::RHI::BlendFactor::AlphaSourceInverse;
-
-            m_depthState.m_enable = false;
-        }
-
-        AZ::RHI::TargetBlendState m_blendState;
-        AZ::RHI::DepthState m_depthState;
-    };
-
-    //! Struct used to pass additional image options.
-    //
-    //! If this is not passed then the defaults are used
-    struct ImageOptions
-    {
-        AZ::Vector3 color = AZ::Vector3(1.0f, 1.0f, 1.0f);
-        Rounding pixelRounding = Rounding::Nearest;
-        bool m_clamp = false;
-        RenderState m_renderState;
-    };
-
-    //! Struct used to pass additional text options - mostly ones that do not change from call to call.
-    //
-    //! If this is not passed then the defaults below are used
-    struct TextOptions
-    {
-        AZStd::string   fontName;               //!< default is "default"
-        unsigned int    effectIndex;            //!< default is 0
-        AZ::Vector3     color;                  //!< default is (1,1,1)
-        HAlign          horizontalAlignment;    //!< default is HAlign::Left
-        VAlign          verticalAlignment;      //!< default is VAlign::Top
-        AZ::Vector2     dropShadowOffset;       //!< default is (0,0), zero offset means no drop shadow is drawn
-        AZ::Color       dropShadowColor;        //!< default is (0,0,0,0), zero alpha means no drop shadow is drawn
-        float           rotation;               //!< default is 0
-        bool            depthTestEnabled;       //!< default is false
-    };
-
-    //! Used to pass in arrays of vertices (e.g. to DrawQuad)
-    struct VertexPosColUV
-    {
-        VertexPosColUV() {}
-        VertexPosColUV(const AZ::Vector2& inPos, const AZ::Color& inColor, const AZ::Vector2& inUV)
-        {
-            position = inPos;
-            color = inColor;
-            uv = inUV;
-        }
-
-        AZ::Vector2         position;   //!< 2D position of vertex
-        AZ::Color           color;      //!< Float color
-        AZ::Vector2         uv;         //!< Texture coordinate
-    };
 
 public: // member functions
 
@@ -116,7 +55,7 @@ public: // member functions
     //! \param imageOptions     Optional struct specifying options that tend to be the same from call to call
     void DrawImage(AZ::Data::Instance<AZ::RPI::Image> image, AZ::Vector2 position, AZ::Vector2 size, float opacity = 1.0f,
         float rotation = 0.0f, const AZ::Vector2* pivotPoint = nullptr, const AZ::Vector2* minMaxTexCoords = nullptr,
-        ImageOptions* imageOptions = nullptr);
+        ImageOptions* imageOptions = nullptr) override;
 
     //! Draw a textured quad where the position specifies the point specified by the alignment.
     //
@@ -135,7 +74,7 @@ public: // member functions
     void DrawImageAligned(AZ::Data::Instance<AZ::RPI::Image> image, AZ::Vector2 position, AZ::Vector2 size,
         HAlign horizontalAlignment, VAlign verticalAlignment,
         float opacity = 1.0f, float rotation = 0.0f, const AZ::Vector2* minMaxTexCoords = nullptr,
-        ImageOptions* imageOptions = nullptr);
+        ImageOptions* imageOptions = nullptr) override;
 
     //! Draw a textured quad where the position, color and uv of each point is specified explicitly
     //
@@ -143,11 +82,11 @@ public: // member functions
     //! \param verts        An array of 4 vertices, in clockwise order (e.g. top left, top right, bottom right, bottom left)
     //! \param pixelRounding Whether and how to round pixel coordinates
     //! \param renderState  Blend mode and depth state
-    virtual void DrawQuad(AZ::Data::Instance<AZ::RPI::Image> image,
+    void DrawQuad(AZ::Data::Instance<AZ::RPI::Image> image,
         VertexPosColUV* verts,
         Rounding pixelRounding = Rounding::Nearest,
         bool clamp = false,
-        const RenderState& renderState = RenderState{});
+        const RenderState& renderState = RenderState{}) override;
 
     //! Draw a line
     //
@@ -156,9 +95,9 @@ public: // member functions
     //! \param color        The color of the line
     //! \param pixelRounding Whether and how to round pixel coordinates
     //! \param renderState  Blend mode and depth state
-    virtual void DrawLine(AZ::Vector2 start, AZ::Vector2 end, AZ::Color color,
+    void DrawLine(AZ::Vector2 start, AZ::Vector2 end, AZ::Color color,
         IDraw2d::Rounding pixelRounding = IDraw2d::Rounding::Nearest,
-        const RenderState& renderState = RenderState{});
+        const RenderState& renderState = RenderState{}) override;
 
     //! Draw a line with a texture so it can be dotted or dashed
     //
@@ -166,10 +105,10 @@ public: // member functions
     //! \param verts        An array of 2 vertices for the start and end points of the line
     //! \param pixelRounding Whether and how to round pixel coordinates
     //! \param renderState  Blend mode and depth state
-    virtual void DrawLineTextured(AZ::Data::Instance<AZ::RPI::Image> image,
+    void DrawLineTextured(AZ::Data::Instance<AZ::RPI::Image> image,
         VertexPosColUV* verts,
         IDraw2d::Rounding pixelRounding = IDraw2d::Rounding::Nearest,
-        const RenderState& renderState = RenderState{});
+        const RenderState& renderState = RenderState{}) override;
     //! Draw a text string. Only supports ASCII text.
     //
     //! The font and effect used to render the text are specified in the textOptions structure
@@ -179,7 +118,7 @@ public: // member functions
     //! \param opacity      The opacity (alpha value) to use to draw the text
     //! \param textOptions  Pointer to an options struct. If null the default options are used
     void DrawText(const char* textString, AZ::Vector2 position, float pointSize,
-        float opacity = 1.0f, TextOptions* textOptions = nullptr);
+        float opacity = 1.0f, TextOptions* textOptions = nullptr) override;
 
     //! Draw a rectangular outline with a texture
     //
@@ -194,43 +133,43 @@ public: // member functions
         AZ::Vector2 rightVec,
         AZ::Vector2 downVec,
         AZ::Color color,
-        uint32_t lineThickness = 0);
+        uint32_t lineThickness = 0) override;
 
     //! Get the width and height (in pixels) that would be used to draw the given text string.
     //
     //! Pass the same parameter values that would be used to draw the string
-    AZ::Vector2 GetTextSize(const char* textString, float pointSize, TextOptions* textOptions = nullptr);
+    AZ::Vector2 GetTextSize(const char* textString, float pointSize, TextOptions* textOptions = nullptr) override;
 
     //! Get the width of the rendering viewport (in pixels).
-    float GetViewportWidth() const;
+    float GetViewportWidth() const override;
 
     //! Get the height of the rendering viewport (in pixels).
-    float GetViewportHeight() const;
+    float GetViewportHeight() const override;
 
     //! Get dpi scale factor
-    float GetViewportDpiScalingFactor() const;
+    float GetViewportDpiScalingFactor() const override;
 
     //! Get the default values that would be used if no image options were passed in
     //
     //! This is a convenient way to initialize the imageOptions struct
-    virtual const ImageOptions& GetDefaultImageOptions() const;
+    const ImageOptions& GetDefaultImageOptions() const override;
 
     //! Get the default values that would be used if no text options were passed in
     //
     //! This is a convenient way to initialize the textOptions struct
-    virtual const TextOptions& GetDefaultTextOptions() const;
+    const TextOptions& GetDefaultTextOptions() const override;
 
     //! Render the primitives that have been deferred
-    void RenderDeferredPrimitives();
+    void RenderDeferredPrimitives() override;
 
     //! Specify whether to defer future primitives or render them right away
-    void SetDeferPrimitives(bool deferPrimitives);
+    void SetDeferPrimitives(bool deferPrimitives) override;
 
     //! Return whether future primitives will be deferred or rendered right away
-    bool GetDeferPrimitives();
+    bool GetDeferPrimitives() override;
 
     //! Set sort key offset for following draws.
-    void SetSortKey(int64_t key);
+    void SetSortKey(int64_t key) override;
 
 private:
 
@@ -377,264 +316,4 @@ protected: // attributes
     AZ::RPI::ViewportContextPtr m_viewportContext;
     AZ::RHI::Ptr<AZ::RPI::DynamicDrawContext> m_dynamicDraw;
     Draw2dShaderData m_shaderData;
-};
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-//! Helper class for using the IDraw2d interface
-//!
-//! The Draw2dHelper class is an inline wrapper that provides the convenience feature of
-//! automatically setting member options structures to their defaults and providing set functions.
-class Draw2dHelper
-{
-public: // member functions
-
-    //! Start a section of 2D drawing function calls that will render to the default viewport
-    Draw2dHelper(bool deferCalls = false)
-    {
-        InitCommon(nullptr, deferCalls);
-    }
-
-    //! Start a section of 2D drawing function calls that will render to the viewport
-    //! associated with the specified Draw2d object
-    Draw2dHelper(CDraw2d* draw2d, bool deferCalls = false)
-    {
-        InitCommon(draw2d, deferCalls);
-    }
-
-    void InitCommon(CDraw2d* draw2d, bool deferCalls)
-    {
-        m_draw2d = draw2d;
-
-        if (!m_draw2d)
-        {
-            // Set to default which is the game's draw 2d object
-            m_draw2d = GetDefaultDraw2d();
-        }
-
-        if (m_draw2d)
-        {
-            m_previousDeferCalls = m_draw2d->GetDeferPrimitives();
-            m_draw2d->SetDeferPrimitives(deferCalls);
-            m_imageOptions = m_draw2d->GetDefaultImageOptions();
-            m_textOptions = m_draw2d->GetDefaultTextOptions();
-        }
-    }
-
-    //! End a section of 2D drawing function calls.
-    ~Draw2dHelper()
-    {
-        if (m_draw2d)
-        {
-            m_draw2d->SetDeferPrimitives(m_previousDeferCalls);
-        }
-    }
-
-    //! Draw a textured quad, optional rotation is counter-clockwise in degrees.
-    //
-    //! See IDraw2d:DrawImage for parameter descriptions
-    void DrawImage(AZ::Data::Instance<AZ::RPI::Image> image, AZ::Vector2 position, AZ::Vector2 size, float opacity = 1.0f,
-        float rotation = 0.0f, const AZ::Vector2* pivotPoint = nullptr, const AZ::Vector2* minMaxTexCoords = nullptr)
-    {
-        if (m_draw2d)
-        {
-            m_draw2d->DrawImage(image, position, size, opacity, rotation, pivotPoint, minMaxTexCoords, &m_imageOptions);
-        }
-    }
-
-    //! Draw a textured quad where the position specifies the point specified by the alignment.
-    //
-    //! See IDraw2d:DrawImageAligned for parameter descriptions
-    void DrawImageAligned(AZ::Data::Instance<AZ::RPI::Image> image, AZ::Vector2 position, AZ::Vector2 size,
-        IDraw2d::HAlign horizontalAlignment, IDraw2d::VAlign verticalAlignment,
-        float opacity = 1.0f, float rotation = 0.0f, const AZ::Vector2* minMaxTexCoords = nullptr)
-    {
-        if (m_draw2d)
-        {
-            m_draw2d->DrawImageAligned(image, position, size, horizontalAlignment, verticalAlignment,
-                opacity, rotation, minMaxTexCoords, &m_imageOptions);
-        }
-    }
-
-    //! Draw a textured quad where the position, color and uv of each point is specified explicitly
-    //
-    //! See IDraw2d:DrawQuad for parameter descriptions
-    void DrawQuad(AZ::Data::Instance<AZ::RPI::Image> image, CDraw2d::VertexPosColUV* verts,
-        IDraw2d::Rounding pixelRounding = IDraw2d::Rounding::Nearest,
-        bool clamp = false,
-        const CDraw2d::RenderState& renderState = CDraw2d::RenderState{})
-    {
-        if (m_draw2d)
-        {
-            m_draw2d->DrawQuad(image, verts, pixelRounding, clamp, renderState);
-        }
-    }
-
-    //! Draw a line
-    //
-    //! See IDraw2d:DrawLine for parameter descriptions
-    void DrawLine(AZ::Vector2 start, AZ::Vector2 end, AZ::Color color,
-        IDraw2d::Rounding pixelRounding = IDraw2d::Rounding::Nearest,
-        const CDraw2d::RenderState& renderState = CDraw2d::RenderState{})
-    {
-        if (m_draw2d)
-        {
-            m_draw2d->DrawLine(start, end, color, pixelRounding, renderState);
-        }
-    }
-
-    //! Draw a line with a texture so it can be dotted or dashed
-    //
-    //! See IDraw2d:DrawLineTextured for parameter descriptions
-    void DrawLineTextured(AZ::Data::Instance<AZ::RPI::Image> image, CDraw2d::VertexPosColUV* verts,
-        IDraw2d::Rounding pixelRounding = IDraw2d::Rounding::Nearest,
-        const CDraw2d::RenderState& renderState = CDraw2d::RenderState{})
-    {
-        if (m_draw2d)
-        {
-            m_draw2d->DrawLineTextured(image, verts, pixelRounding, renderState);
-        }
-    }
-
-    //! Draw a rect outline with a texture
-    //
-    //! See IDraw2d:DrawRectOutlineTextured for parameter descriptions
-    void DrawRectOutlineTextured(AZ::Data::Instance<AZ::RPI::Image> image,
-        UiTransformInterface::RectPoints points,
-        AZ::Vector2 rightVec,
-        AZ::Vector2 downVec,
-        AZ::Color color,
-        uint32_t lineThickness = 0)
-    {
-        if (m_draw2d)
-        {
-            m_draw2d->DrawRectOutlineTextured(image, points, rightVec, downVec, color, lineThickness);
-        }
-    }
-
-    //! Draw a text string. Only supports ASCII text.
-    //
-    //! See IDraw2d:DrawText for parameter descriptions
-    void DrawText(const char* textString, AZ::Vector2 position, float pointSize, float opacity = 1.0f)
-    {
-        if (m_draw2d)
-        {
-            m_draw2d->DrawText(textString, position, pointSize, opacity, &m_textOptions);
-        }
-    }
-
-    //! Get the width and height (in pixels) that would be used to draw the given text string.
-    //
-    //! See IDraw2d:GetTextSize for parameter descriptions
-    AZ::Vector2 GetTextSize(const char* textString, float pointSize)
-    {
-        if (m_draw2d)
-        {
-            return m_draw2d->GetTextSize(textString, pointSize, &m_textOptions);
-        }
-        else
-        {
-            return AZ::Vector2(0, 0);
-        }
-    }
-
-    // State management
-
-    //! Set the blend mode used for images, default is GS_BLSRC_SRCALPHA|GS_BLDST_ONEMINUSSRCALPHA.
-    void SetImageBlendMode(const AZ::RHI::TargetBlendState& blendState) { m_imageOptions.m_renderState.m_blendState = blendState; }
-
-    //! Set the color used for DrawImage and other image drawing.
-    void SetImageColor(AZ::Vector3 color) { m_imageOptions.color = color; }
-
-    //! Set whether images are rounded to have the points on exact pixel boundaries.
-    void SetImagePixelRounding(IDraw2d::Rounding round) { m_imageOptions.pixelRounding = round; }
-
-    //! Set the base state (that blend mode etc is combined with) used for images, default is GS_NODEPTHTEST.
-    void SetImageDepthState(const AZ::RHI::DepthState& depthState) { m_imageOptions.m_renderState.m_depthState = depthState; }
-
-    //! Set image clamp mode
-    void SetImageClamp(bool clamp) { m_imageOptions.m_clamp = clamp; }
-
-    //! Set the text font.
-    void SetTextFont(AZStd::string_view fontName) { m_textOptions.fontName = fontName; }
-
-    //! Set the text font effect index.
-    void SetTextEffectIndex(unsigned int effectIndex) { m_textOptions.effectIndex = effectIndex; }
-
-    //! Set the text color.
-    void SetTextColor(AZ::Vector3 color) { m_textOptions.color = color; }
-
-    //! Set the text alignment.
-    void SetTextAlignment(IDraw2d::HAlign horizontalAlignment, IDraw2d::VAlign verticalAlignment)
-    {
-        m_textOptions.horizontalAlignment = horizontalAlignment;
-        m_textOptions.verticalAlignment = verticalAlignment;
-    }
-
-    //! Set a drop shadow for text drawing. An alpha of zero disables drop shadow.
-    void SetTextDropShadow(AZ::Vector2 offset, AZ::Color color)
-    {
-        m_textOptions.dropShadowOffset = offset;
-        m_textOptions.dropShadowColor = color;
-    }
-
-    //! Set a rotation for the text. The text rotates around its position (taking into account alignment).
-    void SetTextRotation(float rotation)
-    {
-        m_textOptions.rotation = rotation;
-    }
-
-    //! Set wheter to enable depth test for the text
-    void SetTextDepthTestEnabled(bool enabled)
-    {
-        m_textOptions.depthTestEnabled = enabled;
-    }
-
-public: // static member functions
-
-    //! Helper to get the default IDraw2d interface
-    static CDraw2d* GetDefaultDraw2d()
-    {
-        if (gEnv && gEnv->pLyShine) // [LYSHINE_ATOM_TODO][GHI #3569] Remove LyShine global interface pointer from legacy global environment
-        {
-            IDraw2d* draw2d = gEnv->pLyShine->GetDraw2d();
-            return reinterpret_cast<CDraw2d*>(draw2d);
-        }
-
-        return nullptr;
-    }
-
-    //! Round the X and Y coordinates of a point using the given rounding policy
-    template<typename T>
-    static T RoundXY(T value, IDraw2d::Rounding roundingType)
-    {
-        T result = value;
-
-        switch (roundingType)
-        {
-        case IDraw2d::Rounding::None:
-            // nothing to do
-            break;
-        case IDraw2d::Rounding::Nearest:
-            result.SetX(floor(value.GetX() + 0.5f));
-            result.SetY(floor(value.GetY() + 0.5f));
-            break;
-        case IDraw2d::Rounding::Down:
-            result.SetX(floor(value.GetX()));
-            result.SetY(floor(value.GetY()));
-            break;
-        case IDraw2d::Rounding::Up:
-            result.SetX(ceil(value.GetX()));
-            result.SetY(ceil(value.GetY()));
-            break;
-        }
-
-        return result;
-    }
-
-protected: // attributes
-
-    CDraw2d::ImageOptions   m_imageOptions; //!< image options are stored locally and updated by member functions
-    CDraw2d::TextOptions    m_textOptions;  //!< text options are stored locally and updated by member functions
-    CDraw2d* m_draw2d;
-    bool m_previousDeferCalls;
 };
