+++ Gems/LyShine/Code/Source/UiCanvasComponent.h
@@ -292,8 +292,8 @@ public: // member functions
 
     void DebugReportDrawCalls(AZ::IO::HandleType fileHandle, LyShineDebug::DebugInfoDrawCallReport& reportInfo, void* context) const;
 
-    void DebugDisplayElemBounds(CDraw2d* draw2d) const;
-    void DebugDisplayChildElemBounds(CDraw2d* draw2d, const AZ::EntityId entity) const;
+    void DebugDisplayElemBounds(IDraw2d* draw2d) const;
+    void DebugDisplayChildElemBounds(IDraw2d* draw2d, const AZ::EntityId entity) const;
 #endif
 
 public: // static member functions
